Cleanup of index page and preparation for next stage
This commit adds some changes to the database. Indexes were added for
Articles and Ingredients as well as a new table Quantities, which will
hold the complex relationship table of Articles and Ingredients.
Searching for Articles has currently been out commented / disabled because of an
error, that aborded seeding the database. Searching will be reenabled in
a future commit.
The style of the index page has changed and does now conform to the
final version. Some style changes would be very appreciated though as
the page is not really an eye candy.